adretools/requirements.txt
setuptools==69.5.1
wheel==0.43.0
Django==4.2.7
Pillow==10.4.0
pypdf==4.3.1
reportlab==4.0.7
qrcode[pil]==7.4.2
python-barcode==0.15.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
openpyxl==3.1.2
python-pptx==0.6.23
python-docx==1.1.0
cairosvg==2.7.1
requests==2.31.0
python-whois==0.8.0
pandas==2.0.3
opencv-python-headless==4.8.1.78
numpy==1.24.3
google-generativeai==0.3.2
torch==2.0.1
torchvision==0.15.2
transformers==4.35.0